backend/requirements.txt
Flask==0.12
pyjwt==1.4.2
flask-cors==3.0.2
google-auth==0.8.0
requests==2.13.0
requests-toolbelt==0.7.1
GoogleAppEngineCloudStorageClient
